refactor pod sigma proof types to separate submodule
