+++ zk-token-sdk/src/zk_token_elgamal/pod/mod.rs
@@ -1,5 +1,6 @@
 mod elgamal;
 mod pedersen;
+mod sigma_proof;
 
 use {
     crate::zk_token_proof_instruction::ProofType,
@@ -12,6 +13,11 @@ pub use {
     bytemuck::{Pod, Zeroable},
     elgamal::{ElGamalCiphertext, ElGamalPubkey},
     pedersen::{DecryptHandle, PedersenCommitment},
+    sigma_proof::{
+        AggregatedValidityProof, CiphertextCiphertextEqualityProof,
+        CiphertextCommitmentEqualityProof, FeeSigmaProof, PubkeyValidityProof, ValidityProof,
+        ZeroBalanceProof,
+    },
 };
 
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Pod, Zeroable)]
@@ -61,66 +67,6 @@ impl TryFrom<PodProofType> for ProofType {
 #[derive(Clone, Copy, Pod, Zeroable, PartialEq, Eq)]
 #[repr(transparent)]
 pub struct CompressedRistretto(pub [u8; 32]);
-
-/// Serialization of `CiphertextCommitmentEqualityProof`
-#[derive(Clone, Copy)]
-#[repr(transparent)]
-pub struct CiphertextCommitmentEqualityProof(pub [u8; 192]);
-
-// `CiphertextCommitmentEqualityProof` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for CiphertextCommitmentEqualityProof {}
-unsafe impl Pod for CiphertextCommitmentEqualityProof {}
-
-/// Serialization of `CtxtCtxtEqualityProof`
-#[derive(Clone, Copy)]
-#[repr(transparent)]
-pub struct CiphertextCiphertextEqualityProof(pub [u8; 224]);
-
-// `CtxtCtxtEqualityProof` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for CiphertextCiphertextEqualityProof {}
-unsafe impl Pod for CiphertextCiphertextEqualityProof {}
-
-/// Serialization of validity proofs
-#[derive(Clone, Copy)]
-#[repr(transparent)]
-pub struct ValidityProof(pub [u8; 160]);
-
-// `ValidityProof` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for ValidityProof {}
-unsafe impl Pod for ValidityProof {}
-
-/// Serialization of aggregated validity proofs
-#[derive(Clone, Copy)]
-#[repr(transparent)]
-pub struct AggregatedValidityProof(pub [u8; 160]);
-
-// `AggregatedValidityProof` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for AggregatedValidityProof {}
-unsafe impl Pod for AggregatedValidityProof {}
-
-/// Serialization of zero balance proofs
-#[derive(Clone, Copy)]
-#[repr(transparent)]
-pub struct ZeroBalanceProof(pub [u8; 96]);
-
-// `ZeroBalanceProof` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for ZeroBalanceProof {}
-unsafe impl Pod for ZeroBalanceProof {}
-
-/// Serialization of fee sigma proof
-#[derive(Clone, Copy, Pod, Zeroable)]
-#[repr(transparent)]
-pub struct FeeSigmaProof(pub [u8; 256]);
-
-/// Serialization of public-key sigma proof
-#[derive(Clone, Copy, Pod, Zeroable)]
-#[repr(transparent)]
-pub struct PubkeyValidityProof(pub [u8; 64]);
 
 /// Serialization of range proofs for 64-bit numbers (for `Withdraw` instruction)
 #[derive(Clone, Copy)]
